wrapping up refactor pt1
