+++ src/site/client/app/home/verticalNavBar/index.tsx
@@ -1,156 +1,14 @@
-import React, { useEffect, useMemo, useRef, useState } from 'react'
-
-import { ComponentExhibit, Variant, VariantGroup } from '../../../../../api/exhibit/types'
+import React, { EffectCallback, MutableRefObject, useEffect, useRef, useState } from 'react'
+import { ExhibitNode, ExhibitNodeType, PathTree } from '../../../../../api/exhibit/types'
 import { NavLinkKeeyQuery } from '../../../common/navLinkKeepQuery'
-import Ternary from '../../../common/ternary'
 import { useAppSelector } from '../../../store'
 import { LoadingState } from '../../../store/types'
 
-/* This is early PoC code for the vertical nav side bar. It's my first time making one
- * with React so it's not in a high-quality state with concerns separated and such at the moment.
- * TODO: Improve this. See note above.
- *
- * This also should probably be completely rewritten. VSCode doesn't nest their components but
- * rather has a flat list, and they are quite a bit smarter. The state management of these
- * nested components is getting very difficult to work with, and I see now why this is a bad idea.
- */
-
-const VariantEl = (
-  props: {
-    path: string
-    variant: Variant
-  },
-) => (
-  <NavLinkKeeyQuery
-    to={props.path != null ? `${props.path}/${encodeURIComponent(props.variant.name)}` : encodeURIComponent(props.variant.name)}
-    onKeyDown={e => {
-      if (e.key === ' ')
-        (e.target as HTMLElement).click()
-    }}
-  >
-    <i className="far fa-file" />
-    <span className="name">{props.variant.name}</span>
-  </NavLinkKeeyQuery>
-)
-
-const VariantGroupEl = (props: {
-  exhibit: ComponentExhibit<true>
-  variantGroup: VariantGroup
-  path?: string
-  id: string|number
-  isTopLevel?: boolean
-  expandedPaths: string[]
-  onExpandChange: (path: string, newIsExpanded: boolean) => void
-}) => {
-  const isExpandedFromProps = props.expandedPaths.indexOf(props.path) !== -1
-  const isExpanded = useRef(isExpandedFromProps)
-  if (isExpandedFromProps !== isExpanded.current)
-    isExpanded.current = isExpandedFromProps
-
-  const onGroupNameClick = () => {
-    props.onExpandChange(props.path, !isExpanded.current)
-    isExpanded.current = !isExpanded.current
-  }
-
-  return (
-    <div key={props.id} className="variant-group">
-      <button type="button" className="name" onClick={() => onGroupNameClick()}>
-        <i className={`fas ${isExpanded.current ? 'fa-angle-down' : 'fa-angle-right'}`} />
-        <div className="text">{props.variantGroup.name}</div>
-      </button>
-      {isExpanded.current
-        ? (
-          <>
-            <div className="variants">
-              <Ternary
-                bool={(props.isTopLevel ?? true) && (props.exhibit.showDefaultVariant ?? true) && props.exhibit.defaultProps != null}
-                t={(
-                  <VariantEl path={props.path} variant={{ name: 'Default', props: props.exhibit.defaultProps }} key="default" />
-                )}
-              />
-              {Object.values(props.variantGroup.variants).map((variant, i) => (
-                <VariantEl path={props.path} variant={variant} key={i + 1} />
-              ))}
-            </div>
-            <div className="variant-groups">
-              {Object.values(props.variantGroup.variantGroups).map((variantGroup, i) => (
-                <VariantGroupEl
-                  key={i + 1}
-                  id={i + 1}
-                  exhibit={props.exhibit}
-                  variantGroup={variantGroup}
-                  path={props.path != null ? `${props.path}/${encodeURIComponent(variantGroup.name)}` : encodeURIComponent(variantGroup.name)}
-                  isTopLevel={false}
-                  expandedPaths={props.expandedPaths}
-                  onExpandChange={props.onExpandChange}
-                />
-              ))}
-            </div>
-          </>
-        ) : null}
-    </div>
-  )
-}
-
-const ExhibitEl = (props: {
-  exhibit: ComponentExhibit
-  id: string | number
-  expandedPaths: string[]
-  onExpandChange: (path: string, newIsExpanded: boolean) => void
-}) => (
-  props.exhibit.hasProps
-    ? (
-      <VariantGroupEl
-        exhibit={props.exhibit}
-        variantGroup={props.exhibit}
-        path={props.exhibit.name}
-        id={props.id}
-        expandedPaths={props.expandedPaths}
-        onExpandChange={props.onExpandChange}
-      />
-    ) : <VariantEl path="" variant={{ name: props.exhibit.name, props: undefined }} key="default" />
-)
-
-const determineExhibitGroupingInfo = () => {
-  const ungroupedExhibits: ComponentExhibit[] = []
-  // Determine the distinct exhibit group names
-  const groupNameToExhibits: { [groupName: string]: ComponentExhibit[] } = {}
-  const groupNames: string[] = Object.values(exh.default).reduce<string[]>((acc, e) => {
-    if (e.groupName == null) {
-      ungroupedExhibits.push(e)
-      return acc
-    }
-
-    if (groupNameToExhibits[e.groupName] == null)
-      groupNameToExhibits[e.groupName] = []
-
-    groupNameToExhibits[e.groupName].push(e)
-    return acc.indexOf(e.groupName) === -1 ? acc.concat(e.groupName) : acc
-  }, [])
-
-  return {
-    ungroupedExhibits,
-    groupNameToExhibits,
-    groupNames,
-  }
-}
-
-const ExhibitGroupNameEl = (props: {
-  groupName: string
-  isExpanded: boolean
-  onClick: () => void
-}) => (
-  <button type="button" className="name-wrapper" onClick={() => props.onClick()}>
-    <div className="name">
-      <i className={`fas ${props.isExpanded ? 'fa-angle-down' : 'fa-angle-right'}`} />
-      <span className="text">{props.groupName}</span>
-    </div>
-  </button>
-)
+const DEFAULT_WIDTH_PX = 300
 
 type NavBarState = {
-  expandedExhibitGroups: string[]
-  expandedPaths: string[]
+  widthPx: number
+  expandedPaths: { [path: string]: boolean }
 }
 
 const saveNavBarState = (state: NavBarState) => {
@@ -164,208 +22,321 @@ const restoreNavBarState = (): NavBarState => {
     .find(row => row.startsWith('navbar='))
     ?.split('=')[1]
   if (rawValue == null)
-    return { expandedExhibitGroups: [], expandedPaths: [] }
+    return { expandedPaths: {}, widthPx: DEFAULT_WIDTH_PX }
   try {
     const parsed = JSON.parse(rawValue) as NavBarState
+    if (typeof parsed !== 'object')
+      return { expandedPaths: {}, widthPx: DEFAULT_WIDTH_PX }
+
     return {
-      expandedExhibitGroups: [...new Set(parsed.expandedExhibitGroups)] ?? [],
-      expandedPaths: [...new Set(parsed.expandedPaths)] ?? [],
+      expandedPaths: parsed.expandedPaths ?? {},
+      widthPx: parsed.widthPx ?? DEFAULT_WIDTH_PX,
     }
   }
   catch {
-    return { expandedExhibitGroups: [], expandedPaths: [] }
+    return { expandedPaths: {}, widthPx: DEFAULT_WIDTH_PX }
   }
 }
 
-const determineAllPathsOfVariantGroup = (group: VariantGroup, path: string): string[] => {
-  const encodedGroupName = encodeURIComponent(group.name)
-  const thisPath = path != null ? `${path}/${encodedGroupName}` : encodedGroupName
-  return [thisPath]
-    // .concat(Object.values(group.variants).map(v => `${thisPath}/${encodeURIComponent(v.name)}`))
-    .concat(Object.values(group.variantGroups).reduce<string[]>((acc, g) => acc.concat(determineAllPathsOfVariantGroup(g, thisPath)), []))
+const createTopLevelElFocusEffect = (
+  elRef: MutableRefObject<HTMLDivElement>,
+  isElFocusRef: MutableRefObject<boolean>,
+): EffectCallback => () => {
+  const onKeyDown = (e: KeyboardEvent) => {
+    if (e.key === 'Tab') {
+      if (!e.shiftKey && isElFocusRef.current) {
+        const focusableElements: (HTMLAnchorElement | HTMLButtonElement)[] = Array.from(elRef.current.querySelectorAll('a, button'))
+        focusableElements[0]?.focus()
+        e.preventDefault()
+      }
+      elRef.current.classList.remove('focused')
+      isElFocusRef.current = false
+      return
+    }
+
+    const isUp = e.key === 'ArrowUp'
+    const isDown = e.key === 'ArrowDown'
+    if (!isUp && !isDown)
+      return
+
+    if (!(isElFocusRef.current || elRef.current.contains(document.activeElement)))
+      return
+
+    const focusableElements: (HTMLAnchorElement | HTMLButtonElement)[] = Array.from(elRef.current.querySelectorAll('a, button'))
+
+    let focusedElementIndex = 0
+    for (let i = 0; i < focusableElements.length; i += 1) {
+      if (focusableElements[i] === document.activeElement) {
+        focusedElementIndex = i
+        break
+      }
+    }
+
+    const focusedElementIndexChange = isUp && focusedElementIndex > 0
+      ? -1
+      : isDown && focusedElementIndex < focusableElements.length
+        ? 1
+        : 0
+    focusableElements[focusedElementIndex + focusedElementIndexChange]?.focus()
+  }
+
+  const onClick = (e: MouseEvent) => {
+    const isEl = e.target === elRef.current
+    isElFocusRef.current = isEl
+    elRef.current.classList.toggle('focused', isEl)
+  }
+
+  document.addEventListener('click', onClick)
+  document.addEventListener('keydown', onKeyDown)
+  return () => {
+    document.removeEventListener('click', onClick)
+    document.removeEventListener('keydown', onKeyDown)
+  }
 }
 
-const determineAllPaths = () => (
-  Object.values(exh.default)
-    .filter(e => e.hasProps)
-    .reduce<string[]>((acc, exhibit) => (
-      acc.concat(determineAllPathsOfVariantGroup(exhibit as ComponentExhibit<true>, null))
-    ), [])
+const createTopLevelElResizableEffect = (
+  elRef: MutableRefObject<HTMLDivElement>,
+  initialWidthPx: number,
+  onResizeFinish: (newWidthPx: number) => void,
+): EffectCallback => () => {
+  const resizerEl = document.createElement('div')
+  resizerEl.classList.add('exh-resizer')
+  elRef.current.appendChild(resizerEl)
+  elRef.current.style.width = `${initialWidthPx}px`
+
+  const state: any = {
+    initialUserSelectStyleValue: null,
+    initialWidth: 0,
+    x0: 0,
+    x1: 0,
+    mouseDownHandler: null,
+    mouseMoveHandler: null,
+    mouseUpHandler: null,
+  }
+
+  state.mouseDownHandler = (e: MouseEvent) => {
+    state.initialWidth = elRef.current.getBoundingClientRect().width
+    state.x0 = e.screenX
+    state.x1 = e.screenX
+    // Ensure that it's impossible to end up with dupe listeners
+    document.removeEventListener('mouseup', state.mouseUpHandler)
+    document.addEventListener('mouseup', state.mouseUpHandler)
+    document.removeEventListener('mousemove', state.mouseMoveHandler)
+    document.addEventListener('mousemove', state.mouseMoveHandler)
+    state.initialUserSelectStyleValue = elRef.current.style.userSelect
+    elRef.current.style.userSelect = 'none'
+  }
+
+  state.mouseMoveHandler = (e: MouseEvent) => {
+    state.x1 = e.screenX
+    const dx = state.x1 - state.x0
+    elRef.current.style.width = `${state.initialWidth + dx}px`
+  }
+
+  state.mouseUpHandler = (e: MouseEvent) => {
+    state.x1 = e.screenX
+    const dx = state.x1 - state.x0
+    const newWidthPx = state.initialWidth + dx
+    elRef.current.style.width = `${newWidthPx}px`
+
+    elRef.current.style.userSelect = state.initialUserSelectStyleValue
+    document.removeEventListener('mouseup', state.mouseUpHandler)
+    document.removeEventListener('mousemove', state.mouseMoveHandler)
+    onResizeFinish(newWidthPx)
+  }
+
+  resizerEl.addEventListener('mousedown', state.mouseDownHandler)
+  return () => {
+    resizerEl.removeEventListener('mousedown', state.mouseDownHandler)
+    document.removeEventListener('mouseup', state.mouseUpHandler)
+    document.removeEventListener('mousemove', state.mouseMoveHandler)
+    elRef.current.removeChild(resizerEl)
+  }
+}
+
+const indentationPx = 8
+
+const calcNodePaddingLeft = (node: ExhibitNode) => (
+  (indentationPx * (node.pathComponents.length - 1))
 )
 
-export const render = () => {
-  const readyState = useAppSelector(s => s.componentExhibits.ready)
-  const loadingState = useAppSelector(s => s.componentExhibits.loadingState)
+const VariantNodeEl = (props: {
+  node: ExhibitNode<ExhibitNodeType.VARIANT>
+}) => (
+  <NavLinkKeeyQuery
+    to={props.node.path}
+    onKeyDown={e => {
+      if (e.key === ' ')
+        (e.target as HTMLElement).click()
+    }}
+    className="variant"
+    style={{ paddingLeft: calcNodePaddingLeft(props.node) }}
+  >
+    <i className="far fa-file" />
+    <span className="name">{props.node.variant.name}</span>
+  </NavLinkKeeyQuery>
+)
 
+const VariantGroupNodeEl = (props: {
+  node: ExhibitNode<ExhibitNodeType.VARIANT_GROUP>
+  isExpanded: boolean
+  onClick: () => void
+}) => (
+  <button
+    type="button"
+    className="variant-group"
+    style={{ paddingLeft: calcNodePaddingLeft(props.node) }}
+    onClick={props.onClick}
+  >
+    <i className={`fas ${props.isExpanded ? 'fa-angle-down' : 'fa-angle-right'}`} />
+    <div className="text">{props.node.variantGroup.name}</div>
+  </button>
+)
+
+const ExhibitGroupNodeEl = (props: {
+  node: ExhibitNode<ExhibitNodeType.EXHIBIT_GROUP>
+  isExpanded: boolean
+  onClick: () => void
+}) => (
+  <button
+    type="button"
+    className="exhibit-group"
+    style={{ paddingLeft: calcNodePaddingLeft(props.node) }}
+    onClick={props.onClick}
+  >
+    <i className={`fas ${props.isExpanded ? 'fa-angle-down' : 'fa-angle-right'}`} />
+    <div className="text">{props.node.groupName}</div>
+  </button>
+)
+
+const NodeEl = (props: {
+  node: ExhibitNode
+  isExpanded: boolean
+  onClick: () => void
+}) => {
+  if (props.node.type === ExhibitNodeType.EXHIBIT_GROUP)
+    return <ExhibitGroupNodeEl node={props.node} isExpanded={props.isExpanded} onClick={props.onClick} />
+  if (props.node.type === ExhibitNodeType.VARIANT_GROUP)
+    return <VariantGroupNodeEl node={props.node} isExpanded={props.isExpanded} onClick={props.onClick} />
+  if (props.node.type === ExhibitNodeType.VARIANT)
+    return <VariantNodeEl node={props.node} />
+  return null
+}
+
+type GroupTypeNode = ExhibitNode<ExhibitNodeType.EXHIBIT_GROUP | ExhibitNodeType.VARIANT_GROUP>
+
+const PathTreeEl = (props: {
+  pathTree: PathTree
+  expandedPaths: { [path: string]: boolean }
+  onDirExpansionChange: (node: GroupTypeNode, isExpanded: boolean) => void
+}) => (
+  <>
+    {Object.entries(props.pathTree).map(([path, childPathTree]) => {
+      const node = exh.nodes[path]
+      const isExpanded = props.expandedPaths[node.path] === true
+      return (
+        <>
+          <NodeEl node={node} isExpanded={isExpanded} onClick={() => props.onDirExpansionChange(node as GroupTypeNode, !isExpanded)} />
+          {(!isExpanded || typeof childPathTree === 'boolean')
+            ? null
+            : (
+              <PathTreeEl
+                pathTree={childPathTree}
+                expandedPaths={props.expandedPaths}
+                onDirExpansionChange={props.onDirExpansionChange}
+              />
+            )}
+        </>
+      )
+    })}
+  </>
+)
+
+const Render = () => {
   // Restore the nav bar state from cookies once. I think we can just do this in the redux store code instead.
   const hasRestoredNavBarState = useRef(false)
   const initialState = !hasRestoredNavBarState.current ? restoreNavBarState() : null
   hasRestoredNavBarState.current = true
 
-  const [expandedExhibitGroups, setExpandedExhibitGroups] = useState<string[]>(initialState?.expandedExhibitGroups)
-  const [expandedPaths, setExpandedPaths] = useState<string[]>(initialState?.expandedPaths)
-  const exhibitGroupingInfo = useMemo(() => (readyState ? determineExhibitGroupingInfo() : null), [readyState ? exh.default : null])
+  const [expandedPaths, setExpandedPaths] = useState<{ [path: string]: boolean }>(initialState?.expandedPaths)
+  const widthPxRef = useRef(initialState?.widthPx ?? DEFAULT_WIDTH_PX)
+
   const el = useRef<HTMLDivElement>()
   const isElFocus = useRef(false)
 
-  const onGroupNameClick = (groupName: string) => {
-    const currentlyExpanded = expandedExhibitGroups.indexOf(groupName) !== -1
-    const newExpandedExhibitGroups = currentlyExpanded
-      ? expandedExhibitGroups.filter(n => n !== groupName)
-      : expandedExhibitGroups.concat(groupName)
-    setExpandedExhibitGroups(newExpandedExhibitGroups)
+  const onResizeFinish = (newWidthPx: number) => {
     saveNavBarState({
-      expandedExhibitGroups: newExpandedExhibitGroups,
+      widthPx: newWidthPx,
       expandedPaths,
     })
   }
 
-  const onVariantGroupExpandChange = (path: string, newIsExpanded: boolean) => {
-    const isPathExpanded = expandedPaths.indexOf(path) !== -1
-    const newExpandedPaths = newIsExpanded && !isPathExpanded
-      ? expandedPaths.concat(path)
-      : !newIsExpanded && isPathExpanded
-        ? expandedPaths.filter(p => p !== path)
-        : expandedPaths
-    setExpandedPaths(newExpandedPaths)
+  useEffect(createTopLevelElFocusEffect(el, isElFocus), [])
+
+  useEffect(createTopLevelElResizableEffect(el, initialState.widthPx, onResizeFinish), [])
+
+  const onGroupDirExpansionChange = (node: ExhibitNode, newIsExpanded: boolean) => {
+    if (newIsExpanded)
+      expandedPaths[node.path] = true
+    else
+      delete expandedPaths[node.path]
+    setExpandedPaths({ ...expandedPaths })
     saveNavBarState({
-      expandedExhibitGroups,
-      expandedPaths: newExpandedPaths,
+      widthPx: widthPxRef.current,
+      expandedPaths,
     })
   }
 
   const onExpandAllButtonClick = () => {
-    const allPaths = determineAllPaths()
-    setExpandedExhibitGroups(exhibitGroupingInfo.groupNames)
-    setExpandedPaths(allPaths)
+    const newExpandedPaths: { [path: string]: boolean } = {}
+    Object.keys(exh.nodes).forEach(path => newExpandedPaths[path] = true)
+    setExpandedPaths(newExpandedPaths)
     saveNavBarState({
-      expandedExhibitGroups: exhibitGroupingInfo.groupNames,
-      expandedPaths: allPaths,
+      widthPx: widthPxRef.current,
+      expandedPaths: newExpandedPaths,
     })
   }
 
   const onCollapseAllButtonClick = () => {
-    setExpandedExhibitGroups([])
-    setExpandedPaths([])
+    const newExpandedPaths: { [path: string]: boolean } = {}
+    setExpandedPaths(newExpandedPaths)
     saveNavBarState({
-      expandedExhibitGroups: [],
-      expandedPaths: [],
+      widthPx: widthPxRef.current,
+      expandedPaths: newExpandedPaths,
     })
   }
 
-  useEffect(() => {
-    const onKeyDown = (e: KeyboardEvent) => {
-      if (e.key === 'Tab') {
-        if (!e.shiftKey && isElFocus.current) {
-          const focusableElements: (HTMLAnchorElement | HTMLButtonElement)[] = Array.from(el.current.querySelectorAll('a, button'))
-          focusableElements[0]?.focus()
-          e.preventDefault()
-        }
-        el.current.classList.remove('focused')
-        isElFocus.current = false
-        return
-      }
-
-      const isUp = e.key === 'ArrowUp'
-      const isDown = e.key === 'ArrowDown'
-      if (!isUp && !isDown)
-        return
-
-      if (!(isElFocus.current || el.current.contains(document.activeElement)))
-        return
-
-      const focusableElements: (HTMLAnchorElement | HTMLButtonElement)[] = Array.from(el.current.querySelectorAll('a, button'))
-
-      let focusedElementIndex = 0
-      for (let i = 0; i < focusableElements.length; i += 1) {
-        if (focusableElements[i] === document.activeElement) {
-          focusedElementIndex = i
-          break
-        }
-      }
-
-      const focusedElementIndexChange = isUp && focusedElementIndex > 0
-        ? -1
-        : isDown && focusedElementIndex < focusableElements.length
-          ? 1
-          : 0
-      focusableElements[focusedElementIndex + focusedElementIndexChange]?.focus()
-    }
-
-    const onClick = (e: MouseEvent) => {
-      const isEl = e.target === el.current
-      isElFocus.current = isEl
-      el.current.classList.toggle('focused', isEl)
-    }
-
-    document.addEventListener('click', onClick)
-    document.addEventListener('keydown', onKeyDown)
-    return () => {
-      document.removeEventListener('click', onClick)
-      document.removeEventListener('keydown', onKeyDown)
-    }
-  }, [])
-
-  if (readyState) {
-    return (
-      <div
-        className="vertical-nav-bar"
-        ref={el}
-      >
-        <div className="button-bar-1">
-          <button
-            type="button"
-            className="far fa-square-plus"
-            onClick={() => onExpandAllButtonClick()}
-            aria-label="Expand all"
-            title="Expand all"
-          />
-          <button
-            type="button"
-            className="far fa-square-minus"
-            onClick={() => onCollapseAllButtonClick()}
-            aria-label="Collapse all"
-            title="Collapse all"
-          />
-        </div>
-        {exhibitGroupingInfo.groupNames.map((groupName, i) => {
-          const isExpanded = expandedExhibitGroups.indexOf(groupName) !== -1
-          return (
-            <div className={`group ${isExpanded ? 'expanded' : 'collapsed'}`} key={i + 1}>
-              <ExhibitGroupNameEl groupName={groupName} isExpanded={isExpanded} onClick={() => onGroupNameClick(groupName)} />
-              {isExpanded
-                ? (
-                  <div className="exhibits">
-                    {exhibitGroupingInfo.groupNameToExhibits[groupName].map((exhibit, _i) => (
-                      <ExhibitEl
-                        exhibit={exhibit}
-                        key={_i + 1}
-                        id={_i + 1}
-                        expandedPaths={expandedPaths}
-                        onExpandChange={onVariantGroupExpandChange}
-                      />
-                    ))}
-                  </div>
-                ) : null}
-            </div>
-          )
-        })}
-        <div className="ungrouped">
-          <div className="exhibits">
-            {exhibitGroupingInfo.ungroupedExhibits.map((exhibit, i) => (
-              <ExhibitEl
-                exhibit={exhibit}
-                key={i + 1}
-                id={i + 1}
-                expandedPaths={expandedPaths}
-                onExpandChange={onVariantGroupExpandChange}
-              />
-            ))}
-          </div>
-        </div>
+  return (
+    <div className="navigator-side-bar" ref={el}>
+      <div className="button-bar-1">
+        <button
+          type="button"
+          className="far fa-square-plus"
+          onClick={() => onExpandAllButtonClick()}
+          aria-label="Expand all"
+          title="Expand all"
+        />
+        <button
+          type="button"
+          className="far fa-square-minus"
+          onClick={() => onCollapseAllButtonClick()}
+          aria-label="Collapse all"
+          title="Collapse all"
+        />
       </div>
-    )
-  }
+      <div className="nodes">
+        <PathTreeEl pathTree={exh.pathTree} expandedPaths={expandedPaths} onDirExpansionChange={onGroupDirExpansionChange} />
+      </div>
+    </div>
+  )
+}
+
+export const render = () => {
+  const readyState = useAppSelector(s => s.componentExhibits.ready)
+  const loadingState = useAppSelector(s => s.componentExhibits.loadingState)
+
+  if (readyState)
+    return <Render />
 
   if (loadingState === LoadingState.FETCHING)
     return <div className="vertical-nav-bar">Loading exhibits...</div>
