--- src/site/client/assets/styles/home/vertical-nav-bar.scss
@@ -1,185 +1,0 @@
-@import '../common/colors.scss';
-@import '../common/borders.scss';
-
-$indentation: 8px;
-$variant-x-padding: 5px;
-$row-y-sep: 0px;
-$active-color: #ddd;
-$hover-color: #eeeeee;
-
-.row-button {
-  background-color: $color-background;
-  border: none;
-  cursor: pointer;
-  display: block;
-  text-align: left;
-  font-size: 14px;
-  color: lighten($color-text, 15);
-  padding: 0;
-  white-space: nowrap;
-}
-
-@mixin vertical-nav-bar {
-  display: flex;
-  flex-direction: column;
-  flex-shrink: 0;
-  min-width: 150px;
-  border-right: $border;
-  overflow-y: auto;
-
-  .button-bar-1 {
-    display: flex;
-    justify-content: end;
-    padding: 5px;
-
-    button {
-      margin-left: 5px;
-      padding: 5px;
-      cursor: pointer;
-      background-color: $color-background;
-      border: none;
-      border-radius: 5px;
-
-      &:hover {
-        background-color: darken($color-background, 7);
-      }
-    }
-  }
-
-  &.focused {
-    box-shadow: 0 0 0 1px $link-color inset;
-  }
-
-  .group, .ungrouped {
-    padding-left: 5px;
-    padding-right: 5px;
-    display: flex;
-    flex-direction: column;
-    align-items: stretch;
-
-    >.name-wrapper {
-      @extend .row-button;
-      margin-bottom: $row-y-sep;
-      border-bottom: $border;
-      cursor: pointer;
-      padding-bottom: $row-y-sep;
-      user-select: none;
-      padding: 1px;
-
-      &:focus {
-        box-shadow: 0 0 0 1px $link-color inset;
-      }
-
-      .name {
-        padding: 4px 6px; 
-        border-radius: 5px;
-
-        &:hover {
-          background-color: $hover-color;
-        }
-
-        .text {
-          letter-spacing: 2px;
-          color: #777;
-          font-style: italic;
-        }
-
-        i {
-          width: 13px;
-        }
-      }
-    }
-
-    .exhibits {
-      display: flex;
-      flex-direction: column;
-      align-items: stretch;
-      margin-left: $indentation;
-
-      // Special handling for the case of an exhibit with no variants
-      >a {
-        margin-left: $indentation - $variant-x-padding;
-      }
-
-      .variant-group {
-        margin-left: $indentation;
-        display: flex;
-        flex-direction: column;
-        align-items: stretch;
-
-        >.name {
-          @extend .row-button;
-          margin-left: -6px;
-          border-radius: 5px;
-          padding: 3px 5px;
-          margin-bottom: $row-y-sep;
-          cursor: pointer;
-          padding-right: 25px; // Weird workaround for bad interaction with long group names and vertical overflow.
-
-          &:hover {
-            background-color: $hover-color;
-          }
-
-          &:focus {
-            box-shadow: 0 0 0 1px $link-color inset;
-          }
-          
-          i {
-            width: 13px;
-          }
-
-          .text {
-            display: inline-block;
-          }
-        }
-
-        .variants {
-          display: flex;
-          flex-direction: column;
-          margin-left: $indentation;
-
-          >a {
-            margin-left: -$variant-x-padding;
-          }
-        }
-      }
-    }
-  }
-
-  .ungrouped {
-    .exhibits {
-      margin-left: 0;
-      display: flex;
-      flex-direction: column;
-    }
-  }
-
-  a {
-    padding: 1px $variant-x-padding;
-    margin-bottom: $row-y-sep;
-    color: $link-color;
-    text-decoration: none;
-    border-radius: $border-radius;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-
-    &:focus {
-      box-shadow: 0 0 0 1px $link-color inset;
-    }
-
-    &:hover {
-      background-color: $hover-color;
-    }
-
-    &.active {
-      background-color: $active-color;
-      font-weight: bold; // TODO: use workaround to not make boldness cause shift
-    }
-
-    i {
-      margin-right: 6px;
-      font-size: 12px;
-    }
-  }
-}
